+++ src/components/Header/Header.tsx
@@ -47,7 +47,7 @@ export const Header = chakra(({ gaOptions = {}, isProduction = true }) => {
     // key is set to true, try to refresh the accessToken.
     if (data?.data?.statusCode === 401 && data?.data?.expired === true) {
       refreshAccessToken(
-        tokenRefreshLink,
+        tokenRefreshLink(isProduction),
         loginDataCallback,
         deleteCookieValue
       );
@@ -88,7 +88,10 @@ export const Header = chakra(({ gaOptions = {}, isProduction = true }) => {
   }, []);
 
   return (
-    <HeaderComponents.PatronProvider patronName={patronName}>
+    <HeaderComponents.HeaderProvider
+      isProduction={isProduction}
+      patronName={patronName}
+    >
       <Box __css={styles}>
         <SkipNavigation />
         <HeaderComponents.SitewideAlerts />
@@ -122,7 +125,7 @@ export const Header = chakra(({ gaOptions = {}, isProduction = true }) => {
           <HorizontalRule __css={styles.horizontalRule} />
         </header>
       </Box>
-    </HeaderComponents.PatronProvider>
+    </HeaderComponents.HeaderProvider>
   );
 });