+++ CHANGELOG.md
@@ -20,9 +20,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the mobile navigation menu for the `Header` component with `MobileNav` and `MobileNavButton` components.
 - Adds the mobile login menu for the `Header` component.
 - Adds the `useCloseDropDown` hook to manage closing dropdown-like components through clicks outside of it and by pressing the escape key.
-- Adds React context for the `Header`'s patron state data through `PatronContext` and `PatronProvider`.
+- Adds React context for the `Header`'s patron state data through `HeaderContext` and `HeaderProvider`.
 - Fixes `useNYPLBreakpoints` hook so it internally controls its own state.
 - Adds the `react-ga` package to send Google Analytics events.
+- Sets Header Auth links for production and development environments.
 
 ## Prerelease